Remove the unused engine.Increment().
Removed encoding.{Encode,Decode,wrapChecksum,unwrapChecksum}() which
were only used by engine.Increment(). Move encoding.WillOverflow() to
engine.willOverflow() now that there is only a single usage of that
function.
