+++ storage/engine/engine_test.go
@@ -21,7 +21,6 @@ package engine
 import (
 	"bytes"
 	"fmt"
-	"math"
 	"math/rand"
 	"reflect"
 	"sort"
@@ -433,51 +432,6 @@ func TestEngineScan1(t *testing.T) {
 				t.Fatalf("could not run scan: %v", err)
 			}
 			ensureRangeEqual(t, sortedKeys, keyMap, keyvals)
-		}
-	}, t)
-}
-
-func TestEngineIncrement(t *testing.T) {
-	defer leaktest.AfterTest(t)
-	runWithAllEngines(func(engine Engine, t *testing.T) {
-		// Start with increment of an empty key.
-		val, err := Increment(engine, proto.EncodedKey("a"), 1)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if val != 1 {
-			t.Errorf("expected increment to be %d; got %d", 1, val)
-		}
-		// Increment same key by 1.
-		if val, err = Increment(engine, proto.EncodedKey("a"), 1); err != nil {
-			t.Fatal(err)
-		}
-		if val != 2 {
-			t.Errorf("expected increment to be %d; got %d", 2, val)
-		}
-		// Increment same key by 2.
-		if val, err = Increment(engine, proto.EncodedKey("a"), 2); err != nil {
-			t.Fatal(err)
-		}
-		if val != 4 {
-			t.Errorf("expected increment to be %d; got %d", 4, val)
-		}
-		// Decrement same key by -1.
-		if val, err = Increment(engine, proto.EncodedKey("a"), -1); err != nil {
-			t.Fatal(err)
-		}
-		if val != 3 {
-			t.Errorf("expected increment to be %d; got %d", 3, val)
-		}
-		// Increment same key by max int64 value to cause overflow; should return error.
-		if val, err = Increment(engine, proto.EncodedKey("a"), math.MaxInt64); err == nil {
-			t.Error("expected an overflow error")
-		}
-		if val, err = Increment(engine, proto.EncodedKey("a"), 0); err != nil {
-			t.Fatal(err)
-		}
-		if val != 3 {
-			t.Errorf("expected increment to be %d; got %d", 3, val)
 		}
 	}, t)
 }
